fix: setting 컨트롤러 추가
